demo/requirements.txt
flwr==1.9.0
tensorflow-cpu>=2.13,<2.17
numpy>=1.24
